% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oblique_strategies}
\alias{oblique_strategies}
\title{Oblique Strategies}
\format{
An object of class \code{character} of length 222.
}
\usage{
oblique_strategies
}
\description{
Oblique Strategies
}
\details{
A dataset of Brian Eno and Peter Schmidt's Oblique Strategies
}
\keyword{datasets}
