% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vindum_covariates}
\alias{Vindum_covariates}
\title{Covariates for the field Vindum}
\format{
A wrapped \code{SpatRaster} with six layers:
\describe{
\item{bluespot}{Depth of sinks (m)}
\item{ECa}{Apparent electrical conductivity}
\item{elevation}{Elevation (m) above sea level}
\item{mrvbf}{Multiresolution index of valley bottom flatness}
\item{sagawi}{SAGA GIS modified topographic wetness index}
\item{valdepth}{Depth of valleys (m)}
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2019.02.019}
}
\usage{
Vindum_covariates
}
\description{
A wrapped \code{SpatRaster} object with six data layers containing ancillary data from the field Vindum in Denmark (1.6 m resolution), inlcuding variables derived from a digital elevation model and a DUALEM-1 sensor.
}
\examples{
library(terra)
my_cov <- unwrap(Vindum_covariates)
crs(my_cov) <- "EPSG:25832"
}
\keyword{datasets}
