% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vindum_SOM}
\alias{Vindum_SOM}
\title{Soil observatuins for the field Vindum}
\format{
A wrapped \code{SpatVector} with six attributes:
\describe{
\item{ID}{Point ID}
\item{UTMX}{Coordinate x according to ETRS89 UTM Zone 32N (m)}
\item{UTMY}{Coordinate y according to ETRS89 UTM Zone 32N (m)}
\item{SOM}{Measured soil organic matter content (\%)}
\item{logSOM}{Natural log-transformed SOM}
\item{dataset}{Particion into datasets for training and validation}
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2019.02.019}
}
\usage{
Vindum_SOM
}
\description{
A wrapped \code{SpatVector} object containing soil organic matter (SOM) observations from the field Vindum in Denmark.
}
\examples{
library(terra)
my_obs <- unwrap(Vindum_SOM)
crs(my_obs) <- "EPSG:25832"
}
\keyword{datasets}
