% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obliquify.R
\name{obliquify}
\alias{obliquify}
\title{Generate oblique geographic coordinates}
\usage{
obliquify(
  x,
  n_angles = 4,
  angles = NULL,
  na.rm = TRUE,
  n_digits = NULL,
  digits_names = NULL,
  filename = "",
  n_blocks = 4,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{x}{A spatraster object.}

\item{n_angles}{Number of angles for rotation. Defaults to length of angles when angles is not null.}

\item{angles}{Numeric vector with specific angles for rotation, given in radians.}

\item{na.rm}{Exclude areas which are NA in the original SpatRaster.}

\item{n_digits}{Number of digits for rounding the oblique geographic coordinates.}

\item{digits_names}{Number of digits for the angles given in the layer names of the output.}

\item{filename}{File name for writing the output SpatRaster.}

\item{n_blocks}{Number of blocks used for processing the SpatRaster.}

\item{n_cores}{Number of cores used for parallel processing. If NULL, no parallel processing is used.}

\item{...}{Optional arguments for writing the output SpatRaster.}
}
\value{
A SpatRaster with oblique geographic coordinates.
}
\description{
Generate oblique geographic coordinates
}
\examples{
library(terra)
f <- system.file("ex/elev.tif", package = "terra")
r <- rast(f)
ogc_r <- obliquify(r)
}
